Separated time integrator implementation and instantiation
